Began download of raw data and transformation into CSVs
The fetchdata.py script will get the most current version of the MD API data and store it in a "raw_data" sub-directory in it's native format (JSON). Once in the "raw_data", the rest of the code sets up the process to be able to clean the data and store it into the "clean_data" directory as all CSV files. Planning on making a separate python script to hold additional cleaning functions depending on data source and types.